[wasm][debugger][tests] Extract CallFunctionOn tests to a new file
